add checkbox and hovered effect on todos
