+++ src/class/Todo.js
@@ -1,9 +1,10 @@
 export default class Todo {
-  constructor({ title, description, due, priority }) {
+  constructor({ title, due, priority }) {
     this.title = title;
     this.id = crypto.randomUUID();
     this.due = due;
     this.priority = priority;
     this.project = null;
+    this.isCompleted = false;
   }
 }
