+++ src/components/todoList.js
@@ -1,6 +1,7 @@
 import LIBRARY from "../helpers/lib";
 import UI from "../class/UI";
 import PubSub, { events } from "../helpers/PubSub";
+import todoComponent from "./todoComponent";
 
 export default function todoList() {
   const todoList = document.querySelector("#project-view-todo-list");
@@ -11,22 +12,7 @@ export default function todoList() {
   function renderTodo() {
     UI.render(todoList, () => {
       return LIBRARY.getActiveProjectTodo().map((t) => {
-        const list = UI.create("li", {
-          dataset: {
-            id: t.id,
-          },
-          className: t.priority,
-        });
-        const title = UI.create("p", {
-          textContent: t.title,
-          className: "title",
-        });
-        const due = UI.create("p", {
-          textContent: t.due,
-          className: "due",
-        });
-        list.append(title, due);
-        return list;
+        return todoComponent(t);
       });
     });
   }
