+++ src/css/style.css
@@ -70,10 +70,47 @@ main {
           gap: 1rem;
           overflow-y: scroll;
 
-          /* todo container */
+          /* TODO */
           & li {
             border: 1px solid #ccc;
             padding: var(--container-padding);
+            display: flex;
+            align-items: center;
+            gap: 1rem;
+            border-radius: var(--border-radius);
+            background-color: var(--clr-bg-todo);
+            cursor: pointer;
+            position: relative;
+
+            &:hover::after {
+              content: "";
+              border: 2px solid var(--clr-bg-secondary);
+              position: absolute;
+              inset: 0;
+              border-radius: inherit;
+            }
+
+            & .info-container {
+              width: 95%;
+            }
+
+            & .title {
+              font-size: 1.3rem;
+            }
+
+            & .due {
+              color: #777;
+            }
+
+            & .check-box {
+              border: 1px solid #ccc;
+              width: 3rem;
+              height: 3rem;
+              display: flex;
+              justify-content: center;
+              align-items: center;
+              border-radius: var(--border-radius);
+            }
           }
         }
       }
