+++ src/css/vars.css
@@ -2,7 +2,8 @@
   --clr-bg-primary: #fefefe;
   --clr-bg-secondary: #adb2d4;
   --clr-bg-accent: #d1f8ef;
-  --clr-bg-red: #E95C63;
+  --clr-bg-red: #e95c63;
+  --clr-bg-todo: #ecf4f3;
 
   --clr-txt-light: #f6f5f5;
   --clr-txt-dark: #333;
