+++ src/index.js
@@ -11,3 +11,5 @@ projectList();
 todoList();
 createProjectForm();
 createTodoForm();
+
+// TODO: Add checkmarks preview and checkmarks on todos
